# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WA-AKG API Documentation",
    "version": "1.6.4",
    "description": "\n# WhatsApp AI Gateway - Complete API Reference\n\nProfessional WhatsApp Gateway for **Airepro Workflow** (India / IST).\n\n## 🔐 Authentication\n1. **API Key Header**: `X-API-Key: your-api-key`\n2. **Session Cookie**: `next-auth.session-token`\n\n## 🌐 Base URLs\n- Production: `https://workflow.airepro.in/api`\n- API alias: `https://workflow-s.airepro.in/api`\n- Local: `http://localhost:3000/api`\n\n## 📋 Common Parameters\n- **sessionId**: e.g. `mysession-01`\n- **jid** (Personal): `919876543210@s.whatsapp.net` (India +91)\n- **jid** (Group): `120363123456789@g.us`\n\n## 🕒 Timezone\nDefault: **Asia/Kolkata** (IST)\n                "
  },
  "servers": [
    {
      "url": "https://workflow.airepro.in/api",
      "description": "Production — workflow.airepro.in"
    },
    {
      "url": "https://workflow-s.airepro.in/api",
      "description": "Production API alias — workflow-s.airepro.in"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local development"
    },
    {
      "url": "http://127.0.0.1:1730/api",
      "description": "Local Docker / Jenkins loopback (:1730)"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for authentication"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "Session cookie (browser only)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error occurred"
          },
          "error": {
            "type": "string",
            "example": "Detailed error info"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Unauthorized",
              "error": "Unauthorized"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - Access denied",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Forbidden - Cannot access this session",
              "error": "Forbidden - Cannot access this session"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "SessionAuth": []
    }
  ],
  "paths": {
...
    "/scheduler/{sessionId}": {
      "delete": {
        "tags": [
          "Scheduler"
        ],
        "summary": "Delete all scheduled messages for a session",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scheduled messages deleted"
          }
        }
      },
      "get": {
        "tags": [
          "Scheduler"
        ],
        "summary": "List scheduled messages",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of scheduled messages"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Session not found"
          }
        }
      },
      "post": {
        "tags": [
          "Scheduler"
        ],
        "summary": "Create scheduled message",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "jid",
                  "content",
                  "sendAt"
                ],
                "properties": {
                  "jid": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "sendAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "mediaUrl": {
                    "type": "string"
                  },
                  "mediaType": {
                    "type": "string",
                    "enum": [
                      "image",
                      "video",
                      "document"
                    ]
                  }
                }
              },
              "example": {
                "jid": "919876543210@s.whatsapp.net",
                "content": "Reminder: Meeting in 10 mins",
                "sendAt": "2024-12-25T10:00:00.000Z",
                "mediaUrl": "https://example.com/image.jpg",
                "mediaType": "image"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scheduled message created"
          },
          "400": {
            "description": "Missing required fields"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/scheduler/{sessionId}/{scheduleId}": {
      "delete": {
        "tags": [
          "Scheduler"
        ],
        "summary": "Delete scheduled message",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scheduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Message not found"
          }
        }
      },
      "put": {
        "tags": [
          "Scheduler"
        ],
        "summary": "Update scheduled message",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scheduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "jid",
                  "content",
                  "sendAt"
                ],
                "properties": {
                  "jid": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "sendAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "mediaUrl": {
                    "type": "string"
                  },
                  "mediaType": {
                    "type": "string",
                    "enum": [
                      "image",
                      "video",
                      "document"
                    ]
                  }
                }
              },
              "example": {
                "jid": "919876543210@s.whatsapp.net",
                "content": "Updated meeting reminder",
                "sendAt": "2024-12-25T11:00:00.000Z",
                "mediaUrl": "https://example.com/image.jpg",
                "mediaType": "image"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message updated"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Message not found"
          }
        }
      }
    },
...
    "/settings/system": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Update system settings",
        "description": "Update global system configuration (Superadmin/Owner only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "appName": {
                    "type": "string"
                  },
                  "logoUrl": {
                    "type": "string"
                  },
                  "timezone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "appName": {
                      "type": "string"
                    },
                    "logoUrl": {
                      "type": "string"
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "description": "Failed to update settings"
          }
        }
      },
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Get system settings",
        "responses": {
          "200": {
            "description": "System settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "default"
                    },
                    "appName": {
                      "type": "string",
                      "example": "WA-AKG"
                    },
                    "logoUrl": {
                      "type": "string",
                      "example": "https://example.com/logo.png"
                    },
                    "timezone": {
                      "type": "string",
                      "example": "Asia/Kolkata"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}